Skeleton files for creating slides
We start with two files:

- `slides.tex`: an _environment file_ for slides
- `example.tex`: an example to illustrate how to use slides

The first step is to test that everything works. So, we start with skeleton code in both files. Compile the `example.tex` using:

    context example

This will create the following files

- `example.pdf`: the output PDF file
- `example.tuc`: an auxiliary file containing information that is reused across
  multiple runs
- `example.log`: a log file

For now, ignore the `.tuc` and `.log` file.

*Note*: For this example, I use an environment file rather than a module. In
general, modules are setups that are shared across multiple projects while
environments are setups for a specific project. Since the purpose of this
tutorial is to illustrate how to create your own presentation style, I prefer
an environment file over a module.