Skeleton files for creating slides
Start with two files:

- `slides.tex`: an _environment file_ for slides
- `example.tex`: an example to illustrate how to use slides

Start with just skeleton code. Test that everything works:

    context example

This will create the following files

- `example.pdf`: the output PDF file
- `example.tuc`: an auxiliary file containing information that is reused across
  multiple runs
- `example.log`: a log file

Ignore the `.tuc` and `.log` file.

*Note*: For this example, I use an evironment file rather than a module. In
general, modules are setups that are shared across multiple projects while
environments are setups for a specific project. Since the purpose of this
example is to illustrate how to create your own presentation module, I prefer
an evironment file over a module.