Visualize layout
By default, ConTeXt uses large margins, which make sense for paper documents
but not for presentations. I will change the margin in multiple steps. The
first step is just to visualize the margins.

Run: `context example` and notice the green rectangles that show the
text area and the margins. Read

- http://wiki.contextgarden.net/Layout
- http://context.aanhet.net/svn/contextman/context-reference/en/co-pagedesign.pdf

for more details on the layout, and

- http://wiki.contextgarden.net/Command/showframe

for details on the `\showframe` command.