Simple title page
This shows how to create a titlepage. The main idea is to use the `set` key for
variables to call a setup that uses a `makeup` page to create the titlepage.
Notice that the page numbers of the subsequent pages are correct. This is
because we used `\userpage` and `\lastuserpage` to show page numbers, and a
makeup page does not increment the userpage number.

The title page is has no style setup. The next few commits will show how to add
different style elements to the titlepage.